refactor: relocate save_plot_all and fix imports
