--- python/utils/__init__.py
@@ -1,5 +1,0 @@
-"""Utility helpers for the Python pipeline."""
-
-from .state_index_map import state_index_map
-
-__all__ = ["state_index_map"]
