+++ src/plot_fused_vs_truth.py
@@ -7,7 +7,7 @@ from typing import Iterable
 
 import numpy as np
 import matplotlib.pyplot as plt
-from python.utils.save_plot_all import save_plot_all
+from utils.save_plot_all import save_plot_all
 
 
 def plot_fused_vs_truth(
