+++ src/utils/__init__.py
@@ -428,3 +428,20 @@ def interpolate_series(
 
     out = np.vstack([np.interp(t_ref, t_data, series[:, i]) for i in range(series.shape[1])]).T
     return out
+from .state_index_map import state_index_map
+
+__all__ = [
+    "state_index_map",
+    "zero_base_time",
+    "get_data_file",
+    "ensure_dependencies",
+    "detect_static_interval",
+    "compute_C_ECEF_to_NED",
+    "ecef_to_geodetic",
+    "is_static",
+    "save_mat",
+    "normal_gravity",
+    "gravity_ecef",
+    "validate_gravity_vector",
+    "interpolate_series",
+]
